Function toiletcli::parse_flags

pub fn parse_flags(
    args: &Vec<String>,
    flags: &mut [Flag<'_>]
) -> Result<Vec<String>, String>

Parses CLI arguments from sys::env::args.

Short flags are two letter flags starting with one dash (-n). Long flags are flags starting with two dashes (--help). You can combine short SimpleFlag flags, eg. -vAsn will set true to all -v, -A, -s, -n flags. Short flags that use StringFlag can’t be combined.

Returns

Ok

All arguments that are not flags. Changes values passed in enums to matching values from args.

Side effects

SimpleFlag

let value: bool = true;  // if flag is specified.
let value: bool = false; // if not.

StringFlag

let value: String = String::from("<value>"); // if flag is specified AND a value is provided,
let value: String = String::from("");        // if flag is not specified.

Err

On unknown flag;

If StringFlag is specified, but no value is provided for it;

If StringFlag is combined with some other flag.

Example

use std::env::args;
use toiletcli::FlagType;

let mut _args = args();

let mut color = String::new();
let mut show_help = false;

let mut flags = vec![
    (vec!["--colored", "-c"], FlagType::StringFlag(&mut color)),
    (vec!["--help"], FlagType::SimpleFlag(&mut show_help)),
];

let args = toiletcli::parse_flags(&mut _args.collect(), &mut flags);